(move) search field to where it belongs
