+++ src/components/modals/SelectPositionsModal/index.tsx
@@ -7,9 +7,9 @@ import { Button } from 'components/buttons'
 import { ButtonControl, ButtonControlType } from 'components/buttons/ButtonControl'
 import { Modal, ModalProps } from 'components/common/Modal'
 import { TokenIcon } from 'components/common/TokenIcon'
+import { SearchField } from 'components/form/SearchField'
 import { ButtonContainer } from 'components/pureStyledComponents/ButtonContainer'
 import { EmptyContentText } from 'components/pureStyledComponents/EmptyContentText'
-import { SearchField } from 'components/search/SearchField'
 import { InfoCard } from 'components/statusInfo/InfoCard'
 import { InlineLoading } from 'components/statusInfo/InlineLoading'
 import { SpinnerSize } from 'components/statusInfo/common'
