+++ src/pages/ConditionsList/index.tsx
@@ -16,6 +16,7 @@ import { DateFilter } from 'components/filters/DateFilter'
 import { MinMaxFilter } from 'components/filters/MinMaxFilter'
 import { OraclesFilterDropdown } from 'components/filters/OraclesFilterDropdown'
 import { StatusFilterDropdown } from 'components/filters/StatusFilterDropdown'
+import { SearchField } from 'components/form/SearchField'
 import { Switch } from 'components/form/Switch'
 import { EmptyContentText } from 'components/pureStyledComponents/EmptyContentText'
 import { PageTitle } from 'components/pureStyledComponents/PageTitle'
@@ -23,7 +24,6 @@ import { Pill, PillTypes } from 'components/pureStyledComponents/Pill'
 import { Sidebar } from 'components/pureStyledComponents/Sidebar'
 import { SidebarRow } from 'components/pureStyledComponents/SidebarRow'
 import { TwoColumnsCollapsibleLayout } from 'components/pureStyledComponents/TwoColumnsCollapsibleLayout'
-import { SearchField } from 'components/search/SearchField'
 import { InfoCard } from 'components/statusInfo/InfoCard'
 import { InlineLoading } from 'components/statusInfo/InlineLoading'
 import { TableControls } from 'components/table/TableControls'
